+++ readme.md
@@ -22,3 +22,6 @@ Postgres must be installed on localhost. Create a user called search_test, with
 password search_test. The user must be able to create databases.
 
 	go test github.com/IMQS/search/search -db_postgres
+
+# Docker Build
+sudo docker build --build-arg ssh_pvt_key="`cat ~/.ssh/github_rsa`" -t imqs/search:master .
