+++ server/engine_test.go
@@ -1,4 +1,4 @@
-package search
+package server
 
 import (
 	"bytes"
