+++ server/indexer.go
@@ -1,4 +1,4 @@
-package search
+package server
 
 import (
 	//"crypto/sha1"
