+++ server/parser.go
@@ -1,4 +1,4 @@
-package search
+package server
 
 import (
 	"bytes"
